% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Christmas_function.R
\name{Christmas_function}
\alias{Christmas_function}
\title{A Christmas Function}
\usage{
Christmas_function(love = TRUE)
}
\arguments{
\item{love}{Do you Christmas? Defaults to TRUE.}
}
\description{
This function allows you to express your love for christmas.
}
\examples{
Christmas_function()

}
\keyword{christmas}
